+++ services/xsjs/teamPlanHierarchy/level6Service.xsjs
@@ -2,6 +2,7 @@
 $.import("xsplanningtool.services.commonLib","mapper");
 var mapper = $.xsplanningtool.services.commonLib.mapper;
 var httpUtil = mapper.getHttp();
+var hl5 = mapper.getLevel5();
 var hl6 = mapper.getLevel6();
 var ErrorLib = mapper.getErrors();
 var config = mapper.getDataConfig();
@@ -36,15 +37,15 @@ function handleGet(params, userId) {
 
 	}else
 	if(in_hl5_id && !hl5_categories && !hl5_expectedOutcomes){
-
+        hl5.checkPermission(userId, null, in_hl5_id);
 		result = hl6.getHl6ByHl5Id(in_hl5_id);
 	}else
 	if(in_hl6_id){
-
+        hl6.checkPermission(userId, null, in_hl6_id);
 		result = hl6.getHl6ById(in_hl6_id);
 	}else
 	if (param_section && param_section == section){
-		result = hl6.getLevel6ForSearch();
+		result = hl6.getLevel6ForSearch(userId);
 	}else if(hl5_categories && in_hl5_id && hl5_categories == categories){
 		result = hl6.getHl6Categories(in_hl5_id);
 	}else if(hl5_expectedOutcomes && in_hl5_id && hl5_expectedOutcomes == expectedOutcomes) {
@@ -58,6 +59,7 @@ function handleGet(params, userId) {
 //Implementation of PUT call -- Update HL6
 function handlePut(reqBody, userId){
 	var parameters = httpUtil.getUrlParameters();
+    hl6.checkPermission(userId, null, parameters.get('HL6_ID') || reqBody.hl6.in_hl6_id);
 	if(parameters.length > 0){
 		var aCmd = parameters.get('method');
 		var hl6Id = parameters.get('HL6_ID');
@@ -84,12 +86,14 @@ function handlePut(reqBody, userId){
 
 //Implementation of DELETE call -- Delete HL6
 function handleDelete(reqBody, userId){
+    hl6.checkPermission(userId, null, reqBody.in_hl6_id);
 	var result =  hl6.deleteHl6(reqBody, userId);
 	return httpUtil.handleResponse(result,httpUtil.OK,httpUtil.AppJson);
 }
 
 //Implementation of POST call -- Insert HL6
-function handlePost(reqBody, userId) {	
+function handlePost(reqBody, userId) {
+    hl5.checkPermission(userId, null, reqBody.hl6.in_hl5_id);
 	var result = hl6.insertHl6(reqBody, userId); //return new L6 Id
 	return httpUtil.handleResponse(result,httpUtil.OK,httpUtil.AppJson);
 }
